requirements.txt
fastapi==0.116.0
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.42
psycopg[binary]==3.2.10
python-multipart==0.0.20
pydantic-settings==2.10.1
httpx==0.28.1
beautifulsoup4==4.13.4
PyPDF2==3.0.1
python-docx==1.2.0
jinja2==3.1.6
python-dotenv==1.1.1
